+++ lib/autokey/iomediator/iomediator.py
@@ -16,12 +16,13 @@
 import threading
 import time
 import queue
+import os
 
 import autokey
 from autokey import common
 from autokey.configmanager.configmanager import ConfigManager
 from autokey.configmanager.configmanager_constants import INTERFACE_TYPE
-from autokey.interface import XRecordInterface, AtSpiInterface
+from autokey.gnome_interface import GnomeExtensionWindowInterface
 from autokey.sys_interface.clipboard import Clipboard
 from autokey.model.phrase import SendMode
 
@@ -55,21 +56,34 @@ class IoMediator(threading.Thread):
         self.app = service.app
         
         # Modifier tracking
-        self.modifiers = {
-            Key.CONTROL: False,
-            Key.ALT: False,
-            Key.ALT_GR: False,
-            Key.SHIFT: False,
-            Key.SUPER: False,
-            Key.HYPER: False,
-            Key.META: False,
-            Key.CAPSLOCK: False,
-            Key.NUMLOCK: False
-        }
+        self.modifiers = {}
+        for key in MODIFIERS:
+            self.modifiers[key]=False
+
+        # self.interfaceType="uinput"
+        session_type = os.environ.get("XDG_SESSION_TYPE")
+        if session_type == "wayland":
+            self.interfaceType = "uinput"
+        elif session_type == "x11":
+            pass
+        elif session_type is None:
+            pass
         
-        if self.interfaceType == X_RECORD_INTERFACE:
+        if self.interfaceType == "uinput":
+            self.windowInterface = GnomeExtensionWindowInterface()
+        else:
+            from autokey.interface import XWindowInterface
+            self.windowInterface = XWindowInterface()
+
+
+        if self.interfaceType == "uinput":
+            from autokey.uinput_interface import UInputInterface
+            self.interface = UInputInterface(self, self.app)
+        elif self.interfaceType == X_RECORD_INTERFACE:
+            from autokey.interface import XRecordInterface
             self.interface = XRecordInterface(self, self.app)
         else:
+            from autokey.interface import AtSpiInterface
             self.interface = AtSpiInterface(self, self.app)
 
         self.clipboard = Clipboard()
@@ -111,7 +125,9 @@ class IoMediator(threading.Thread):
     
     def handle_modifier_down(self, modifier):
         """
-        Updates the state of the given modifier key to 'pressed'
+        Updates the state of the given modifier key to 'pressed'.
+
+        :param modifier: Should be AutoKey Key value
         """
         logger.debug("%s pressed", modifier)
         if modifier in (Key.CAPSLOCK, Key.NUMLOCK):
@@ -152,6 +168,7 @@ class IoMediator(threading.Thread):
 
             # We make a copy here because the wait_for... functions modify the listeners,
             # and we want this processing cycle to complete before changing what happens
+            logger.debug("Raw Key: {} | Modifiers: {} | Key: {} | Window Info: {}".format(raw_key, modifiers, key, window_info))
             for target in self.listeners.copy():
                 target.handle_keypress(raw_key, modifiers, key, window_info)
 
@@ -184,6 +201,7 @@ class IoMediator(threading.Thread):
     @staticmethod
     def _send_string(string, interface):
         modifiers = []
+        logger.debug("Sending string sections: %s", KEY_SPLIT_RE.split(string))
         for section in KEY_SPLIT_RE.split(string):
             if len(section) > 0:
                 if Key.is_key(section[:-1]) and section[-1] == '+' and section[:-1] in MODIFIERS:
@@ -237,7 +255,7 @@ class IoMediator(threading.Thread):
                 backspaces += 1
             else:
                 backspaces += len(section)
-                
+        logger.debug("Sending backspaces: %d", backspaces)
         self.send_backspace(backspaces)
 
     def send_key(self, key_name):
@@ -249,6 +267,7 @@ class IoMediator(threading.Thread):
         self.interface.fake_keydown(key_name)
 
     def release_key(self, key_name):
+        logger.debug("Release key: %s", key_name)
         key_name = key_name.replace('\n', "<enter>")
         self.interface.fake_keyup(key_name)